Get LaTeX code for the connection bivariate GMM features (mu and sigma)

Get LaTeX code for the connection bivariate GMM features (mu and sigma)

Usage

print_connection_models_table(GMM, label, caption, full_width, path = NULL)

Arguments

GMM

Gaussian Mixture Models obtained from function get_connection_models

label

character, e.g. "tab:gmm"

caption

character, table caption

full_width

logical, if true the "*" will be added next to the "table" tag

path

character, file path to write the latex table to. Is must have ".tex" extension. If it is NULL, then the character string is returned instead of writing a file.

Value

character, LaTeX code

Examples

# The package evprof provides example objects of connection and energy
# Gaussian Mixture Models obtained from California's open data set
# (see California article in package website) created with functions
# `get_connection models` and `get_energy models`.

# Get the working days connection models
connection_models <- evprof::california_GMM$workdays$connection_models

# Print connection GMM tables
print_connection_models_table(
  GMM = connection_models,
  label = "tab:con-gmm",
  caption = "Connection GMM",
  full_width = TRUE
)
#> [1] "\\begin{table*}\n\\resizebox{\\linewidth}{!} {\n\\begin{tabular}{l|c|c|c}\n\\hline\nUser profile & Centroid ($\\mu$) & Covariance ($\\Sigma$) & Share (\\%) \\\\\n\\hline\n\\multirow{4}{*}{Visit}& $\\begin{array}{cc} 2.837981 \\\\ 0.66478 \\end{array}$&$\\begin{array}{cc} 0.014548 & 0.00479 \\\\ 0.00479 & 0.366194 \\end{array}$&18\\\\ \\cline{2-4} & $\\begin{array}{cc} 2.59185 \\\\ 1.382431 \\end{array}$&$\\begin{array}{cc} 0.01905 & -0.048368 \\\\ -0.048368 & 0.261887 \\end{array}$&36\\\\ \\cline{2-4} & $\\begin{array}{cc} 1.963827 \\\\ 1.759812 \\end{array}$&$\\begin{array}{cc} 0.030619 & -0.044756 \\\\ -0.044756 & 0.161827 \\end{array}$&28\\\\ \\cline{2-4} & $\\begin{array}{cc} 2.306782 \\\\ 0.617851 \\end{array}$&$\\begin{array}{cc} 0.111771 & -0.123963 \\\\ -0.123963 & 0.816233 \\end{array}$&18 \\\\ \\hline\n\\multirow{3}{*}{Worktime}& $\\begin{array}{cc} 1.988369 \\\\ 2.260861 \\end{array}$&$\\begin{array}{cc} 0.001819 & -0.00138 \\\\ -0.00138 & 0.009553 \\end{array}$&29\\\\ \\cline{2-4} & $\\begin{array}{cc} 1.887171 \\\\ 2.268155 \\end{array}$&$\\begin{array}{cc} 0.009629 & -0.001792 \\\\ -0.001792 & 0.006395 \\end{array}$&34\\\\ \\cline{2-4} & $\\begin{array}{cc} 2.2242 \\\\ 2.119886 \\end{array}$&$\\begin{array}{cc} 0.013847 & -0.015517 \\\\ -0.015517 & 0.038308 \\end{array}$&37 \\\\ \\hline\n\\end{tabular}}\n\\caption{\\label{tab:con-gmm}Connection GMM}\n\\end{table*}"