Clustered EV charging sessions example

Example of an charging sessions data set that has been clustered by evprof functions. (see this article).

Usage

california_ev_sessions_profiles

Format

california_ev_sessions_profiles

A tibble object with standard variable names defined by evprof

Source

https://ev.caltech.edu/dataset