Density plot in 3D, considering Start time and Connection duration as variables

Density plot in 3D, considering Start time and Connection duration as variables

Usage

plot_density_3D(
  sessions,
  start = getOption("evprof.start.hour"),
  eye = list(x = -1.5, y = -1.5, z = 1.5),
  log = FALSE
)

Arguments

sessions

tibble, sessions data set in evprof standard format.

start

integer, start hour in the x axis of the plot. This is only used when log = FALSE.

eye

list containing x, y and z points of view. Example: list(x = -1.5, y = -1.5, z = 1.5)

log

logical, whether to transform ConnectionStartDateTime and ConnectionHours variables to natural logarithmic scale (base = exp(1)).

Value

plotly plot (html)

Examples

plot_density_3D(california_ev_sessions, start = 3)